+++ src/components/common/PageWrapper.tsx
@@ -2,14 +2,12 @@ import React from 'react';
 
 import * as Styled from './TextPageWrapper.style';
 
-export enum PageWidth {
-  ULTRA_NARROW = 'ULTRA_NARROW',
-  NARROW = 'NARROW',
-  REGULAR = 'REGULAR',
-  WIDE = 'WIDE',
-}
+export { PageWidth } from './TextPageWrapper.style';
 
-const PageWrapper: React.FC<{ pageWidth?: PageWidth }> = ({ children, pageWidth = PageWidth.REGULAR }) => (
+const PageWrapper: React.FC<{ pageWidth?: Styled.PageWidth }> = ({
+  children,
+  pageWidth = Styled.PageWidth.REGULAR,
+}) => (
   <Styled.CenteredContainer>
     <Styled.Centered pageWidth={pageWidth}>{children}</Styled.Centered>
   </Styled.CenteredContainer>
